+++ src/main/kotlin/io/hndrs/annotation/processing/repository/ModelHelper.kt
@@ -12,9 +12,9 @@ import kotlin.reflect.KClass
 object ModelHelper {
 
     fun getPackageName(element: Element): String {
-        var foundPackage: Boolean = true
+        var foundPackage: Boolean = false
         var nextElement: Element = element
-        lateinit var packageName: String
+        var packageName: String? = null
         do {
             val enclosingElement = nextElement.enclosingElement
             if (enclosingElement is PackageElement) {
@@ -24,15 +24,19 @@ object ModelHelper {
                 nextElement = enclosingElement
             }
         } while (!foundPackage)
-        return packageName
+        if (packageName != null && !packageName.isBlank()) {
+            return packageName
+        }
+        throw RepositoryGeneratorException("Could not resolve package name for $element")
     }
 
+    @Throws(RepositoryGeneratorException::class)
     fun getIdTypeName(element: Element): String {
-        return element.enclosedElements.first {
+        return element.enclosedElements.firstOrNull {
             it.getAnnotation(Id::class.java) != null
-        }.let {
+        }?.let {
             getClass(it).simpleName.toString()
-        }
+        } ?: throw RepositoryGeneratorException("Could not find @Id annotated property in ${element.simpleName}")
     }
 
     fun getParameterMetas(element: Element): List<ParameterMeta> {