+++ src/test/kotlin/io/hndrs/test/SimpleEntity.kt
@@ -1,0 +1,10 @@
+package io.hndrs.test
+
+import io.hndrs.annotation.processing.repository.GenerateRepository
+import org.springframework.data.annotation.Id
+
+@GenerateRepository(type = GenerateRepository.Type.MONGO)
+data class SimpleEntity(
+    @Id
+    val id: String
+)